Remove AbstractCommCareCase
