+++ corehq/form_processor/models/__init__.py
@@ -24,15 +24,12 @@ from corehq.sql_db.models import PartitionedModel
 from corehq.util.json import CommCareJSONEncoder
 from corehq.util.models import TruncatingCharField
 
-from .abstract import (
-    DEFAULT_PARENT_IDENTIFIER,
-    AbstractCommCareCase,
-)
+from .abstract import DEFAULT_PARENT_IDENTIFIER
 from ..exceptions import AttachmentNotFound, UnknownActionType
 from ..track_related import TrackRelatedChanges
 from .attachment import Attachment, AttachmentContent, AttachmentMixin  # noqa: F401
 from .forms import XFormInstance, XFormOperation  # noqa: F401
-from .mixin import IsImageMixin, SaveStateMixin
+from .mixin import CaseToXMLMixin, IsImageMixin, SaveStateMixin
 
 
 STANDARD_CHARFIELD_LENGTH = 255
@@ -48,7 +45,7 @@ CaseAction = namedtuple("CaseAction", ["action_type", "updated_known_properties"
 
 
 class CommCareCase(PartitionedModel, models.Model, RedisLockableMixIn,
-                   AttachmentMixin, AbstractCommCareCase, TrackRelatedChanges,
+                   AttachmentMixin, CaseToXMLMixin, TrackRelatedChanges,
                    MessagingCaseContactMixin):
     DOC_TYPE = 'CommCareCase'
     partition_attr = 'case_id'
@@ -145,6 +142,19 @@ class CommCareCase(PartitionedModel, models.Model, RedisLockableMixIn,
 
     def dynamic_case_properties(self):
         return OrderedDict(sorted(self.case_json.items()))
+
+    def get_properties_in_api_format(self):
+        return dict(list(self.dynamic_case_properties().items()) + list({
+            "external_id": self.external_id,
+            "owner_id": self.owner_id,
+            # renamed
+            "case_name": self.name,
+            # renamed
+            "case_type": self.type,
+            # renamed
+            "date_opened": self.opened_on,
+            # all custom properties go here
+        }.items()))
 
     def to_api_json(self, lite=False):
         from ..serializers import CommCareCaseAPISerializer
@@ -368,6 +378,90 @@ class CommCareCase(PartitionedModel, models.Model, RedisLockableMixIn,
     def on_tracked_models_cleared(self, model_class=None):
         self._saved_indices.reset_cache(self)
 
+    def resolve_case_property(self, property_name):
+        """
+        Takes a case property expression and resolves the necessary references
+        to get the case property value(s).
+
+        property_name - The case property expression. Examples: name, parent/name,
+                        parent/parent/name
+
+        Returns a list of named tuples of (case, value), where value is the
+        resolved case property value and case is the case that yielded that value.
+        There can be more than one tuple in the returned result if a case has more
+        than one parent or grandparent.
+        """
+        result = []
+        self._resolve_case_property(property_name, result)
+        return result
+
+    def _resolve_case_property(self, property_name, result):
+        CasePropertyResult = namedtuple('CasePropertyResult', 'case value')
+
+        if property_name.lower().startswith('parent/'):
+            parents = self.get_parent(identifier=DEFAULT_PARENT_IDENTIFIER)
+            for parent in parents:
+                parent._resolve_case_property(property_name[7:], result)
+            return
+
+        if property_name.lower().startswith('host/'):
+            host = self.host
+            if host:
+                host._resolve_case_property(property_name[5:], result)
+            return
+
+        if property_name == '_id':
+            property_name = 'case_id'
+
+        result.append(CasePropertyResult(
+            self,
+            self.get_case_property(property_name)
+        ))
+
+    @memoized
+    def get_index_map(self, reversed=False):
+        indices = self.indices if not reversed else self.reverse_indices
+        return get_index_map(indices)
+
+    @memoized
+    def get_attachment_map(self):
+        return {
+            name: {
+                'url': self.get_attachment_server_url(att.identifier),
+                'content_type': att.content_type,
+            }
+            for name, att in self.case_attachments.items()
+        }
+
+    def to_xml(self, version, include_case_on_closed=False):
+        from lxml import etree as ElementTree
+        from casexml.apps.phone.xml import get_case_element
+        if self.closed:
+            if include_case_on_closed:
+                elem = get_case_element(self, ('create', 'update', 'close'), version)
+            else:
+                elem = get_case_element(self, ('close'), version)
+        else:
+            elem = get_case_element(self, ('create', 'update'), version)
+        return ElementTree.tostring(elem, encoding='utf-8')
+
+    def get_attachment_server_url(self, name):
+        """
+        A server specific URL for remote clients to access case attachment resources async.
+        """
+        if name in self.case_attachments:
+            from dimagi.utils import web
+            from django.urls import reverse
+            return "%s%s" % (
+                web.get_url_base(),
+                reverse("api_case_attachment", kwargs={
+                    "domain": self.domain,
+                    "case_id": self.case_id,
+                    "attachment_id": name,
+                }),
+            )
+        return None
+
     @classmethod
     def get_obj_id(cls, obj):
         return obj.case_id
@@ -427,6 +521,16 @@ class CommCareCase(PartitionedModel, models.Model, RedisLockableMixIn,
         ]
         app_label = "form_processor"
         db_table = CommCareCase_DB_TABLE
+
+
+def get_index_map(indices):
+    return dict([
+        (index.identifier, {
+            "case_type": index.referenced_type,
+            "case_id": index.referenced_id,
+            "relationship": index.relationship,
+        }) for index in indices
+    ])
 
 
 class CaseAttachment(PartitionedModel, models.Model, SaveStateMixin, IsImageMixin):
